made corrections to task 3, 5
